finished the types.ts file and update planning.ts to weekly-planning.ts
